+++ custom_components/metlink_explorer/sensor.py
@@ -204,7 +204,7 @@ class MetlinkRouteSensor(CoordinatorEntity, SensorEntity):
             "name": f"{transportation_name} Route {route_short_name}",
             "manufacturer": "Metlink",
             "model": transportation_name,
-            "sw_version": "0.4.6",
+            "sw_version": "0.4.7",
         }
 
     @property
@@ -298,7 +298,7 @@ class MetlinkDirectionSensor(CoordinatorEntity, SensorEntity):
             "name": f"{transportation_name} Route {route_short_name}",
             "manufacturer": "Metlink",
             "model": transportation_name,
-            "sw_version": "0.4.6",
+            "sw_version": "0.4.7",
         }
 
     @property